requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
attrs==25.3.0
    # via
    #   service-identity
    #   twisted
autobahn==24.4.2
    # via daphne
automat==25.4.16
    # via twisted
cffi==1.17.1
    # via cryptography
channels==4.2.2
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==4.2.1
    # via -r requirements.in
constantly==23.10.4
    # via twisted
cryptography==44.0.2
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==4.1.2
    # via -r requirements.in
django==4.2.20
    # via
    #   -r requirements.in
    #   channels
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-environ==0.12.0
    # via -r requirements.in
django-filter==25.1
    # via -r requirements.in
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.0
    # via -r requirements.in
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.10
    # via
    #   hyperlink
    #   twisted
incremental==24.7.2
    # via twisted
msgpack==1.1.0
    # via channels-redis
psycopg[binary]==3.2.6
    # via -r requirements.in
psycopg-binary==3.2.6
    # via psycopg
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycparser==2.22
    # via cffi
pyjwt==2.9.0
    # via djangorestframework-simplejwt
pyopenssl==25.0.0
    # via twisted
redis==5.2.1
    # via channels-redis
service-identity==24.2.0
    # via twisted
sqlparse==0.5.3
    # via django
twisted[tls]==24.11.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.13.2
    # via
    #   psycopg
    #   pyopenssl
    #   twisted
zope-interface==7.2
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
